==> 00.2_create_pgs_job_scripts.R <==
#00.2_create_pgs_job_scripts.R

#load required packages
library(genotools)
library(purrr)

# Major depression (MDD)
mdd2019 <- list(user="p471-jaaskelu",
                host="p471-hpc-01.tsd.usit.no",
                software="ldpred2",
                shortname="mdd2019",
                account="p471",
                cpu_time="05:00:00",
                memory="8G",
                cores=32,
                lib="//ess/p471/data/durable/common/rlibs/4-2-1",
                sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
                outputs_dir="//ess/p471/data/durable/common/pgs_directory",
                sumstats_filename = "mdd2019_withNeff.txt",
                genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
                genotype_data="genoHapMap3plus_N200k.rds",
                ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
                case_control = TRUE,
                A1 = "A1",
                A2 = "A2",
                stat = "LogOR",
                stat_type = "OR",
                pvalue = "P",
                SNP = "MarkerName",
                CHR = "CHR",
                BP = "BP",
                SE = "StdErrLogOR",
                N = "Neff",
                MAF = "MAF",
                maf_threshold = 0.01,
                pmid = "30718901") %>%
  as.data.frame()

# Attention-deficit/hyperactivity disorder (ADHD)
adhd <- list(user="p471-jaaskelu",
             host="p471-hpc-01.tsd.usit.no",
             software="ldpred2",
             shortname="adhd",
             account="p471",
             cpu_time="05:00:00",
             memory="8G",
             cores=32,
             lib="//ess/p471/data/durable/common/rlibs/4-2-1",
             sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
             outputs_dir="//ess/p471/data/durable/common/pgs_directory",
             sumstats_filename = "adhd_withNeff.txt",
             genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
             genotype_data="genoHapMap3plus_N200k.rds",
             ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
             case_control = TRUE,
             A1 = "A1",
             A2 = "A2",
             stat = "OR",
             stat_type = "OR",
             pvalue = "P",
             SNP = "SNP",
             CHR = "CHR",
             BP = "BP",
             SE = "SE",
             N = "Neff",
             info = "INFO",
             maf_threshold = 0.01,
             pmid = "30478444") %>%
  as.data.frame()

# Autism
asd <- list(user="p471-jaaskelu",
            host="p471-hpc-01.tsd.usit.no",
            software="ldpred2",
            shortname="asd",
            account="p471",
            cpu_time="05:00:00",
            memory="8G",
            cores=32,
            lib="//ess/p471/data/durable/common/rlibs/4-2-1",
            sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
            outputs_dir="//ess/p471/data/durable/common/pgs_directory",
            sumstats_filename = "asd_withNeff.txt",
            genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
            genotype_data="genoHapMap3plus_N200k.rds",
            ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
            case_control = TRUE,
            A1 = "A1",
            A2 = "A2",
            stat = "OR",
            stat_type = "OR",
            pvalue = "P",
            SNP = "SNP",
            CHR = "CHR",
            BP = "BP",
            SE = "StdErr",
            N = "Neff",
            info = "INFO",
            MAF = "maf_1000G",
            maf_threshold = 0.01,
            pmid = "30804558") %>%
  as.data.frame()

# Anorexia nervosa
an2019 <- list(user="p471-jaaskelu",
               host="p471-hpc-01.tsd.usit.no",
               software="ldpred2",
               shortname="an2019",
               account="p471",
               cpu_time="05:00:00",
               memory="8G",
               cores=32,
               lib="//ess/p471/data/durable/common/rlibs/4-2-1",
               sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
               outputs_dir="//ess/p471/data/durable/common/pgs_directory",
               sumstats_filename = "an2019_withNeff.txt",
               genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
               genotype_data="genoHapMap3plus_N200k.rds",
               ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
               case_control = TRUE,
               A1 = "ALT",
               A2 = "REF",
               stat = "BETA",
               stat_type = "beta",
               pvalue = "PVAL",
               SNP = "ID",
               CHR = "CHROM",
               BP = "POS",
               SE = "SE",
               N = "Neff",
               maf_threshold = 0.01,
               pmid = "31308545") %>%
  as.data.frame()
 
# Anxiety
anx <- list(user="p471-jaaskelu",
            host="p471-hpc-01.tsd.usit.no",
            software="ldpred2",
            shortname="anx",
            account="p471",
            cpu_time="05:00:00",
            memory="8G",
            cores=32,
            lib="//ess/p471/data/durable/common/rlibs/4-2-1",
            sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
            outputs_dir="//ess/p471/data/durable/common/pgs_directory",
            sumstats_filename = "anx_withNeff.txt",
            genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
            genotype_data="genoHapMap3plus_N200k.rds",
            ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
            case_control = TRUE,
            A1 = "Allele1",
            A2 = "Allele2",
            stat = "Effect",
            stat_type = "beta",
            pvalue = "P",
            SNP = "SNP",
            CHR = "CHR",
            BP = "BP",
            SE = "StdErr",
            N = "Neff",
            MAF = "maf_1000G",
            maf_threshold = 0.01,
            pmid = "31748690") %>%
  as.data.frame()

# Schizophrenia
scz2022 <- list(user="p471-jaaskelu",
                host="p471-hpc-01.tsd.usit.no",
                software="ldpred2",
                shortname="scz2022",
                account="p471",
                cpu_time="05:00:00",
                memory="8G",
                cores=32,
                lib="//ess/p471/data/durable/common/rlibs/4-2-1",
                sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
                outputs_dir="//ess/p471/data/durable/common/pgs_directory",
                sumstats_filename = "scz2022_withNeff.txt",
                genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
                genotype_data="genoHapMap3plus_N200k.rds",
                ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
                case_control = TRUE,
                A1 = "A1",
                A2 = "A2",
                stat = "BETA",
                stat_type = "beta",
                pvalue = "PVAL",
                SNP = "SNP",
                CHR = "CHROM",
                BP = "POS",
                SE = "StdErr",
                N = "Neff",
                info = "IMPINFO",
                MAF = "maf_1000G",
                maf_threshold = 0.01,
                pmid = "35396580") %>%
  as.data.frame()

# Tourette's syndrome (TS)
ts <- list(user="p471-jaaskelu",
           host="p471-hpc-01.tsd.usit.no",
           software="ldpred2",
           shortname="ts",
           account="p471",
           cpu_time="05:00:00",
           memory="8G",
           cores=32,
           lib="//ess/p471/data/durable/common/rlibs/4-2-1",
           sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
           outputs_dir="//ess/p471/data/durable/common/pgs_directory",
           sumstats_filename = "ts_withNeff.txt",
           genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
           genotype_data="genoHapMap3plus_N200k.rds",
           ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
           case_control = TRUE,
           A1 = "A1",
           A2 = "A2",
           stat = "OR",
           stat_type = "OR",
           pvalue = "P",
           SNP = "SNP",
           CHR = "CHR",
           BP = "BP",
           SE = "SE",
           N = "Neff",
           info = "INFO",
           maf_threshold = 0.01,
           pmid = "30818990") %>%
  as.data.frame()

# Obsessive compulsive disorder (OCD)
ocd <- list(user="p471-jaaskelu",
            host="p471-hpc-01.tsd.usit.no",
            software="ldpred2",
            shortname="ocd",
            account="p471",
            cpu_time="05:00:00",
            memory="8G",
            cores=32,
            lib="//ess/p471/data/durable/common/rlibs/4-2-1",
            sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
            outputs_dir="//ess/p471/data/durable/common/pgs_directory",
            sumstats_filename = "ocd_withNeff.txt",
            genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
            genotype_data="genoHapMap3plus_N200k.rds",
            ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
            case_control = TRUE,
            A1 = "A1",
            A2 = "A2",
            stat = "OR",
            stat_type = "OR",
            pvalue = "P",
            SNP = "SNP",
            CHR = "CHR",
            BP = "BP",
            SE = "SE",
            N = "Neff",
            info = "INFO",
            maf_threshold = 0.01,
            pmid = "28761083") %>%
  as.data.frame()

# Post-traumatic stress disorder (PTSD)
ptsd <- list(user="p471-jaaskelu",
             host="p471-hpc-01.tsd.usit.no",
             software="ldpred2",
             shortname="ptsd",
             account="p471",
             cpu_time="05:00:00",
             memory="8G",
             cores=32,
             lib="//ess/p471/data/durable/common/rlibs/4-2-1",
             sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
             outputs_dir="//ess/p471/data/durable/common/pgs_directory",
             sumstats_filename = "ptsd_withNeff.txt",
             genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
             genotype_data="genoHapMap3plus_N200k.rds",
             ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
             case_control = TRUE,
             A1 = "Allele1",
             A2 = "Allele2",
             stat = "Effect",
             stat_type = "beta",
             pvalue = "P-value",
             SNP = "MarkerName",
             CHR = "CHR",
             BP = "BP",
             SE = "StdErr",
             N = "Neff",
             maf_threshold = 0.01,
             pmid = "28439101") %>%
  as.data.frame()

# Bipolar disorder
bip2021 <- list(user="p471-jaaskelu",
                host="p471-hpc-01.tsd.usit.no",
                software="ldpred2",
                shortname="bip2021",
                account="p471",
                cpu_time="05:00:00",
                memory="8G",
                cores=32,
                lib="//ess/p471/data/durable/common/rlibs/4-2-1",
                sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
                outputs_dir="//ess/p471/data/durable/common/pgs_directory",
                sumstats_filename = "bip2021_withNeff.txt",
                genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
                genotype_data="genoHapMap3plus_N200k.rds",
                ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
                case_control = TRUE,
                A1 = "A1",
                A2 = "A2",
                stat = "BETA",
                stat_type = "beta",
                pvalue = "PVAL",
                SNP = "SNP",
                CHR = "CHR",
                BP = "BP",
                SE = "SE",
                N = "Neff",
                maf_threshold = 0.01,
                pmid = "34002096") %>%
  as.data.frame()

# Alcohol dependence
alcdep2018eur <- list(user="p471-jaaskelu",
                      host="p471-hpc-01.tsd.usit.no",
                      software="ldpred2",
                      shortname="alcdep2018eur",
                      account="p471",
                      cpu_time="05:00:00",
                      memory="8G",
                      cores=32,
                      lib="//ess/p471/data/durable/common/rlibs/4-2-1",
                      sumstats_dir="//ess/p471/cluster/p/p471/cluster/common/gwas_sumstats/",
                      outputs_dir="//ess/p471/data/durable/common/pgs_directory",
                      sumstats_filename = "alcdep2018eur_withNeff.txt",
                      genotype_dir="//ess/p471/data/durable/data/genetic_data/MoBaPsychGen_v1/",
                      genotype_data="genoHapMap3plus_N200k.rds",
                      ldref_dir="//ess/p471/data/durable/data/genetic/ldref/",
                      case_control = TRUE,
                      A1 = "A1",
                      A2 = "A2",
                      stat = "OR",
                      stat_type = "OR",
                      pvalue = "P",
                      SNP = "SNP",
                      CHR = "CHR",
                      BP = "BP",
                      SE = "SE",
                      N = "Neff",
                      info = "INFO",
                      maf_threshold = 0.01,
                      pmid = "30482948") %>%
  as.data.frame()

# List of data frames
df_list <- list(adhd,asd,scz2022,bip2021,mdd2019,anx,alcdep2018eur,ptsd,ocd,an2019,ts)

# Merge dataframes
merged_df <- df_list %>%
  purrr::reduce(full_join, by = NULL)

# Save out
save(merged_df, file="./data/sstats/ldpred2_inputs.RData")
