==> 03_genetic_correlations.R <==
#03_genetic_correlations.R

#Genetic correlations between the intercept of differentiation and total 
#problems with 11 psychiatric and neurodevelopmental conditions are estimated in 
#LDSC with the scripts located in the folder 'shell scripts/genetic correlations'.
