==> 04_run_structural_models.R <==
#04_run_structural_models.R

#This script runs the correlated factor, hierarchical, and bifactor models with
#differentiation and total problems predicting the latent factors, by sourcing 
#the script: '04.1_specify_structural_models.R'. It also tabulates the results,
#whereas plots are generated by the script: '04.2_plot_structural_models.R'.

#install genomicSEM and dependencies
#install.packages("N:/durable/projects/differentiation_genomics/packages/rlang_1.0.2.tar",
#                  repos = NULL,
#                  type = "source")
#devtools::install_local("N:/durable/common/software/GenomicSEM_0.0.5.zip")

#load necessary packages
library(GenomicSEM)
library(tidyverse)
library(flextable)
library(officer)

#load variance/covariance matrix of genomics associated with differentiation 
#and total scores, and liability to 11 psychiatric conditions
load(file="./data/cov_wpsych.RData")

#define common settings for all models
covstruc<-cov_wpsych
estimation<-"DWLS"
std.lv=FALSE

#specify structural models by sourcing:
source("./scripts/04.1_specify_structural_models.R")

#################################
### basic latent growth model ###
#################################

Model_basic <- usermodel(covstruc=covstruc,
                         model=model_basic,
                         estimation=estimation,
                         std.lv=std.lv)
#model fit
Model_basic$modelfit #Good fit.

#results
Model_basic$results #Virtually no variation in the slope of diff and tot.
                    #Therefore, specify an intercept-only model.

Model_no_slope <- usermodel(covstruc=covstruc,
                            model=model_no_slope,
                            estimation=estimation,
                            std.lv=std.lv)
#model fit
Model_no_slope$modelfit #Little decrement in fit when dropping the slope.

#results
Model_no_slope$results #Significant variance in intercept of diff and tot.

################################
### correlated factor models ###
################################

#run model with differentiation predicting correlated factors
ModelD_4factors <- usermodel(covstruc=covstruc,
                             model=modelD_4factors,
                             estimation=estimation,
                             std.lv = std.lv)
#model fit
ModelD_4factors$modelfit #good fit

#filter results
tmpD4factors <- ModelD_4factors$results %>%
  filter(lhs%in%c("i","int","psy","neu","com"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("int","psy","neu","com"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "diff_4factors")

#print results
tmpD4factors

#save results
save(tmpD4factors, file="./output/tmpD4factors.RData")

#run model with differentiation predicting correlated factors, but with 
#independent paths to the indicators of the "neu" factor (not converging)
ModelD_4factors_ind <- usermodel(covstruc=covstruc,
                                 model=modelD_4factors_ind,
                                 estimation=estimation,
                                 std.lv = std.lv)

#run model with total problems predicting correlated factors 
ModelT_4factors <- usermodel(covstruc=covstruc,
                             model=modelT_4factors,
                             estimation=estimation,
                             std.lv=std.lv)
#print the model fit
ModelT_4factors$modelfit #good fit

#filter results
tmpT4factors <- ModelT_4factors$results %>%
  filter(lhs%in%c("i","int","psy","neu","com"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("int","psy","neu","com"))|lhs==rhs  ,
         !lhs=="i") %>%
  mutate(model = "tot_4factors")

#print results
tmpT4factors

#save results
save(tmpT4factors, file="./output/tmpT4factors.RData")

#run model with diff + total problems predicting correlated factors 
ModelDT_4factors <- usermodel(covstruc=covstruc,
                              model=modelDT_4factors,
                              estimation=estimation,
                              std.lv=std.lv)
#print the model fit
ModelDT_4factors$modelfit #good fit

#filter results
tmpDT4factors <- ModelDT_4factors$results %>%
  filter(lhs%in%c("i1","i2","int","psy","neu","com"),
         op == "~"|op=="~~",
         (rhs=="i1"|rhs=="i2"&lhs%in%c("int","psy","neu","com"))|lhs==rhs) %>%
  mutate(model = "both_4factors")

#print results
tmpDT4factors

#save results
save(tmpDT4factors, file="./output/tmpDT4factors.RData")

## create table correlated factor model results
load(file="./output/tmpD4factors.RData")
load(file="./output/tmpT4factors.RData")

#differentiation - correlated factors
estsD_cor_all <- tmpD4factors %>% 
  mutate(Unstand_SE = as.numeric(Unstand_SE),
         STD_Genotype_SE = as.numeric(STD_Genotype_SE),
         across(where(is.numeric), ~ round(., 3))) %>%
  filter(op != "~~") %>%
  mutate(fdr_pval = p.adjust(p_value, method = "fdr")) %>%
  select(lhs,model,STD_All,fdr_pval) %>%
  `colnames<-`(c("outcome","model","STD_All","fdr_pval")) %>%
  mutate(model = "Differentiation")

# total problems - correlated factors
estsT_cor_all <- tmpT4factors %>% 
  mutate(Unstand_SE = as.numeric(Unstand_SE),
         STD_Genotype_SE = as.numeric(STD_Genotype_SE),
         across(where(is.numeric), ~ round(., 3))) %>%
  filter(op != "~~") %>%
  mutate(fdr_pval = p.adjust(p_value, method = "fdr")) %>%
  select(lhs,model,STD_All,fdr_pval) %>%
  `colnames<-`(c("outcome","model","STD_All","fdr_pval")) %>%
  mutate(model = "Total problems")

# merge diff and tot
ests_cor_both <- estsD_cor_all %>%
  full_join(estsT_cor_all) %>%
  mutate(model=as.factor(model))

df <- data.frame(
  model = rep(c("Total problems", "Differentiation"), each = 4),
  outcome = rep(c("int", "com", "neu", "psy"), times = 2))

# create table correlated factor model
tbl_cor_both <- ests_cor_both %>%
  left_join(df, by = c("model", "outcome")) %>%
  mutate(across(where(is.numeric), round, digits=2)) %>%
  select(model,outcome,STD_All,fdr_pval) %>%
  mutate(model = if_else(model == "Differentiation", "d", "t")) %>%
  pivot_wider(names_from = model, 
              values_from = c(STD_All,fdr_pval),
              names_sep = ".") %>%
  select(outcome, ends_with(".d"), ends_with(".t")) %>%
  mutate(outcome= as.factor(outcome))

levels(tbl_cor_both$outcome)[levels(tbl_cor_both$outcome)=="com"] <- "Compulsive"
levels(tbl_cor_both$outcome)[levels(tbl_cor_both$outcome)=="psy"] <- "Psychotic"
levels(tbl_cor_both$outcome)[levels(tbl_cor_both$outcome)=="neu"] <- "Neurodevelopmental"
levels(tbl_cor_both$outcome)[levels(tbl_cor_both$outcome)=="int"] <- "Internalising"

(ft_tbl_cor_both <- tbl_cor_both %>% 
    flextable() %>% 
    set_header_labels(values=list(outcome="Latent factor",
                                  STD_All.d="STD_EST",FDR_pval.d="",
                                  STD_All.t="STD_EST",FDR_pval.t="")) %>%
    add_header_row(values=list("","Differentiation","Total problems"), colwidths = c(1,2,2)) %>%
    compose(i = c(2,2), j = c(3,5), part = "header", value = as_paragraph("P",as_sub("FDR"))) %>%
    flextable::font(fontname = "Arial", part= "all") %>% 
    theme_box() %>%
    border_outer(border=fp_border(color="gray80",width=2),part="all") %>%
    border_inner(border=fp_border(color='gray80',width=1),part="all") %>%
    hline(i=2,border=fp_border(color="gray50",width=3),part="header") %>%
    bg(bg="#F1F1F1",part="header") %>% 
    align(align = "center", part="all") %>% 
    set_table_properties(layout = "autofit"))

save_as_docx(ft_tbl_cor_both, path = "./tables/Table_S7_correlated_factor_output.docx")

##################################
### hierarchical factor models ###
##################################

#run hierarchical model with differentiation predicting p-factor 
ModelD_hi_p <- usermodel(covstruc=covstruc,
                         model=modelD_hier_p,
                         estimation=estimation,
                         std.lv=std.lv)
#print the model fit
ModelD_hi_p$modelfit #acceptable fit

#filter results
diff_hier_p <- ModelD_hi_p$results %>%
  filter(rhs%in%c("i","p"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("p"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "diff_hier_p")

#print results
diff_hier_p

#save/load results
save(diff_hier_p, file="./output/diff_hier_p.RData")
load(file="./output/diff_hier_p.RData")

#run model with sub-factors from hierarchical model as outcomes
ModelD_hi_sub <- usermodel(covstruc=covstruc,
                           model=modelD_hi_sub,
                           estimation=estimation,
                           std.lv=std.lv)
#print the model fit
ModelD_hi_sub$modelfit #good fit

#filter results
diff_hier_sub <- ModelD_hi_sub$results %>%
  filter(rhs%in%c("i","int","psy","neu","com"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("int","psy","neu","com"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "diff_hier_sub")

#print results
diff_hier_sub

#save/load results
save(diff_hier_sub, file="./output/diff_hier_sub.RData")
load(file="./output/diff_hier_sub.RData")

#calculate the chi-square and save it in a vector called Q_chisq
Q_chisq<-ModelD_hi_p[[1]]$chisq-ModelD_hi_sub[[1]]$chisq

#calculate the df associated with this chi-square statistic and name it df
df<-ModelD_hi_p[[1]]$df[1]-ModelD_hi_sub[[1]]$df[1]

#calculate the p-value associated with the Q-statistic, using the relevant df. 
#This saves the p-values in a vector called Q_chisq_pval 
Q_chisq_pval_p_D<-pchisq(Q_chisq,df,lower.tail=FALSE)

cbind(ModelD_hi_p[[1]],Q_chisq_pval_p_D)

#run model of differentiation and total problems combined, 
#with p-factor from hierarchical model as predictor
ModelDT_hi_p <- usermodel(covstruc=covstruc,
                          model=modelDT_hi_p,
                          estimation=estimation,
                          std.lv=std.lv)
#print the model fit
ModelDT_hi_p$modelfit

#filter results
both_hi_p <- ModelDT_hi_p$results %>%
  filter(rhs%in%c("i1","i2","p"),
         op == "~"|op=="~~",
         (rhs=="i1"|rhs=="i2"&lhs%in%c("p"))|lhs==rhs) %>%
  mutate(model = "both_hier_p")

#print results
both_hi_p

#save/load results
save(both_hi_p, file="./output/both_hi_p.RData")
load(file="./output/both_hi_p.RData")

#run model of differentiation and total problems combined, 
#with sub-factors from hierarchical model as outcomes
ModelDT_hi_sub <- usermodel(covstruc=covstruc,
                            model=modelDT_hi_sub,
                            estimation=estimation,
                            std.lv=std.lv)
#print the model fit
ModelDT_hi_sub$modelfit

#filter results
both_hi_sub <- ModelDT_hi_sub$results %>%
  filter(rhs%in%c("i1","i2","com","psy","neu","int"),
         op == "~"|op=="~~",
         (rhs=="i1"|rhs=="i2"&lhs%in%c("com","psy","neu","int"))|lhs==rhs) %>%
  mutate(model = "both_hier_sub")

#print results
both_hi_sub

#save/load results
save(both_hi_sub, file="./output/both_hi_sub.RData")
load(file="./output/both_hi_sub.RData")

#calculate the chi-square and save it in a vector called Q_chisq
Q_chisq<-ModelDT_hi_p[[1]]$chisq-ModelDT_hi_sub[[1]]$chisq

#calculate the df associated with this chi-square statistic and name it df
df<-ModelDT_hi_p[[1]]$df[1]-ModelDT_hi_sub[[1]]$df[1]

#calculate the p-value associated with the Q-statistic, using the relevant df. 
#This saves the p-values in a vector called Q_chisq_pval 
Q_chisq_pval<-pchisq(Q_chisq,df,lower.tail=FALSE)

cbind(ModelDT_hi_p[[1]],Q_chisq_pval)

#run model for total problems, with p-factor from hierarchical model as predictor
ModelT_hi_p <- usermodel(covstruc=covstruc,
                         model=modelT_hi_p,
                         estimation=estimation,
                         std.lv=std.lv)
#print the model fit
ModelT_hi_p$modelfit #acceptable fit

#filter results
tot_hier_p <- ModelT_hi_p$results %>%
  filter(rhs%in%c("i","p"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("p"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "tot_hier_p")

#print results
tot_hier_p

#save/load results
save(tot_hier_p, file="./output/tot_hier_p.RData")
load(file="./output/tot_hier_p.RData")

#run model for total problems with sub-factors from hierarchical model as outcomes
ModelT_hi_sub <- usermodel(covstruc=covstruc,
                           model=modelT_hi_sub,
                           estimation=estimation,
                           std.lv=std.lv)
#print the model fit
ModelT_hi_sub$modelfit #acceptable fit

#filter results
tot_hier_sub <- ModelT_hi_sub$results %>%
  filter(lhs%in%c("i","com","psy","neu","int"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("com","psy","neu","int"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "tot_hier_sub")

#print results
tot_hier_sub

#save/load results
save(tot_hier_sub, file="./output/tot_hier_sub.RData")
load(file="./output/tot_hier_sub.RData")

#calculate the chi-square and save it in a vector called Q_chisq
Q_chisq<-ModelT_hi_p[[1]]$chisq-ModelT_hi_sub[[1]]$chisq

#calculate the df associated with this chi-square statistic and name it df
df<-ModelT_hi_p[[1]]$df[1]-ModelT_hi_sub[[1]]$df[1]

#calculate the p-value associated with the Q-statistic, using the relevant df. 
#This saves the p-values in a vector called Q_chisq_pval 
Q_chisq_pval_p_T<-pchisq(Q_chisq,df,lower.tail=FALSE)

cbind(ModelT_hi_p[[1]],Q_chisq_pval_p_T)

# create table hierarchical model

# load results
load(file="./output/diff_hier_p.RData")
load(file="./output/diff_hier_sub.RData")
load(file="./output/tot_hier_p.RData")
load(file="./output/tot_hier_sub.RData")

#merge
diff_hi_all <- diff_hier_p %>%
  full_join(diff_hier_sub)

tot_hi_all <- tot_hier_p %>%
  full_join(tot_hier_sub)

#differentiation - hierarchical
estsD_hi_all <- diff_hi_all %>% 
  mutate(Unstand_SE = as.numeric(Unstand_SE),
         STD_Genotype_SE = as.numeric(STD_Genotype_SE),
         across(where(is.numeric), ~ round(., 3))) %>%
  filter(rhs == "i" & op != "~~") %>%
  mutate(fdr_pval = p.adjust(p_value, method = "fdr")) %>%
  select(lhs,model,STD_All,fdr_pval) %>%
  `colnames<-`(c("outcome","model","STD_All","fdr_pval")) %>%
  mutate(model = "Differentiation")

#total - hierarchical
estsT_hi_all <- tot_hi_all %>% 
  mutate(Unstand_SE = as.numeric(Unstand_SE),
         STD_Genotype_SE = as.numeric(STD_Genotype_SE),
         across(where(is.numeric), ~ round(., 3))) %>%
  filter(rhs == "i" & op != "~~") %>%
  mutate(fdr_pval = p.adjust(p_value, method = "fdr")) %>%
  select(lhs,model,STD_All,fdr_pval) %>%
  `colnames<-`(c("outcome","model","STD_All","fdr_pval"))  %>%
  mutate(model = "Total problems")

# merge diff and tot
ests_hi_both <- estsD_hi_all %>%
  full_join(estsT_hi_all) %>%
  mutate(model=as.factor(model))

# create dataframe to add Q_pval
df_hi <- data.frame(
  model = rep(c("Total problems", "Differentiation"), each = 1),
  outcome = rep(c("p"), times = 2))

# create table hierarchical factor model
tbl_hi_both <- ests_hi_both %>%
  left_join(df_hi, by = c("model", "outcome")) %>%
  mutate(across(where(is.numeric), round, digits=2)) %>%
  select(model,outcome,STD_All,fdr_pval) %>%
  mutate(model = if_else(model == "Differentiation", "d", "t")) %>%
  pivot_wider(names_from = model, 
              values_from = c(STD_All,fdr_pval),
              names_sep = ".") %>%
  select(outcome, ends_with(".d"), ends_with(".t")) %>%
  mutate(outcome=as.factor(outcome))

levels(tbl_hi_both$outcome)[levels(tbl_hi_both$outcome)=="p"] <- "P-factor"
levels(tbl_hi_both$outcome)[levels(tbl_hi_both$outcome)=="com"] <- "Compulsive"
levels(tbl_hi_both$outcome)[levels(tbl_hi_both$outcome)=="psy"] <- "Psychotic"
levels(tbl_hi_both$outcome)[levels(tbl_hi_both$outcome)=="neu"] <- "Neurodevelopmental"
levels(tbl_hi_both$outcome)[levels(tbl_hi_both$outcome)=="int"] <- "Internalising"

(ft_tbl_hi_both <- tbl_hi_both %>% 
    flextable() %>% 
    set_header_labels(values=list(rhs="Latent factor",
                                  STD_All.d="STD_EST",FDR_pval.d="",
                                  STD_All.t="STD_EST",FDR_pval.t="")) %>%
    add_header_row(values=list("","Differentiation","Total problems"), colwidths = c(1,2,2)) %>%
    compose(i = c(2,2), j = c(3,5), part = "header", value = as_paragraph("P",as_sub("FDR"))) %>%
    flextable::font(fontname = "Arial", part= "all") %>% 
    theme_box() %>%
    border_outer(border=fp_border(color="gray80",width=2),part="all") %>%
    border_inner(border=fp_border(color='gray80',width=1),part="all") %>%
    hline(i=2,border=fp_border(color="gray50",width=3),part="header") %>%
    bg(bg="#F1F1F1",part="header") %>% 
    align(align = "center", part="all") %>% 
    set_table_properties(layout = "autofit"))

save_as_docx(ft_tbl_hi_both, path = "./tables/Table_S8_hierarchical_model_output.docx")

#######################
### bifactor models ###
#######################

#run model for diff with p from the bifactor model as the outcome
ModelD_bi_p <- usermodel(covstruc=covstruc,
                         model=modelD_bi_p,
                         estimation=estimation,
                         std.lv=std.lv)

#print the model fit
ModelD_bi_p$modelfit #acceptable fit

#filter results
diff_bi_p <- ModelD_bi_p$results %>%
  filter(lhs%in%c("i","p"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("p"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "diff_bi_p")

#print results
diff_bi_p

#save results
save(diff_bi_p, file="./output/diff_bi_p.RData")

#run model for diff with sub-factors from bifactor model as outcomes
ModelD_bi_sub <- usermodel(covstruc=covstruc,
                           model=modelD_bi_sub,
                           estimation=estimation,
                           std.lv=std.lv)
#print the model fit
ModelD_bi_sub$modelfit #good fit

#filter results
diff_bi_sub <- ModelD_bi_sub$results %>%
  filter(lhs%in%c("i","com","psy","neu","int"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("com","psy","neu","int"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "diff_bi_sub")

#print results
diff_bi_sub

#save results
save(diff_bi_sub, file="./output/diff_bi_sub.RData")

#run model for tot with p from the bifactor model as the outcome
ModelT_bi_p <- usermodel(covstruc=covstruc,
                         model=modelT_bi_p,
                         estimation=estimation,
                         std.lv=std.lv)

#print the model fit
ModelT_bi_p$modelfit #acceptable fit

#filter results
tot_bi_p <- ModelT_bi_p$results %>%
  filter(lhs%in%c("i","p"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("p"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "tot_bi_p")

#print results
tot_bi_p

#save results
save(tot_bi_p, file="./output/tot_bi_p.RData")

#run model for tot w/sub-factors from bifactor as outcomes
ModelT_bi_sub<-usermodel(covstruc=covstruc,
                         model=modelT_bi_sub,
                         estimation=estimation,
                         std.lv=std.lv)

#print the model fit
ModelT_bi_sub$modelfit #good fit

#filter results
tot_bi_sub <- ModelT_bi_sub$results %>%
  filter(lhs%in%c("i","com","psy","neu","int"),
         op == "~"|op=="~~",
         (rhs=="i"&lhs%in%c("com","psy","neu","int"))|lhs==rhs,
         !lhs=="i") %>%
  mutate(model = "tot_bi_sub")

#print results
tot_bi_sub

#save results
save(tot_bi_sub, file="./output/tot_bi_sub.RData")

#run model for diff+tot with p from bifactor model a the outcome
ModelDT_bi_p <- usermodel(covstruc=covstruc,
                          model=modelDT_bi_p,
                          estimation=estimation,
                          std.lv=std.lv)

ModelDT_bi_p$modelfit #good fit

#run model for diff+tot with p and sub-factors from bifactor model as outcomes,
#to facilitate a Q-trait model fit comparison to assess the role of the p-factor
ModelDT_bi_all <- usermodel(covstruc=covstruc,
                            model=modelDT_bi_all,
                            estimation=estimation,
                            std.lv=std.lv)

ModelDT_bi_all$modelfit #good fit

#calculate the chi-square and save it in a vector called Q_chisq
Q_chisq<-ModelDT_bi_p[[1]]$chisq-ModelDT_bi_all[[1]]$chisq

#calculate the df associated with this chi-square statistic and name it df
df<-ModelDT_bi_p[[1]]$df[1]-ModelDT_bi_all[[1]]$df[1]

#calculate the p-value associated with the Q-statistic, using the relevant df. 
#This saves the p-values in a vector called Q_chisq_pval 
Q_chisq_pval_p_DT<-pchisq(Q_chisq,df,lower.tail=FALSE)

#print
Q_chisq_pval_p_DT

# create table bifactor model

# load results
load(file="./output/diff_bi_p.RData")
load(file="./output/diff_bi_sub.RData")
load(file="./output/tot_bi_p.RData")
load(file="./output/tot_bi_sub.RData")

#merge
diff_bi_all <- diff_bi_p %>%
  full_join(diff_bi_sub)

tot_bi_all <- tot_bi_p %>%
  full_join(tot_bi_sub)

#differentiation - bifactor
estsD_bi_all <- diff_bi_all %>% 
  mutate(Unstand_SE = as.numeric(Unstand_SE),
         STD_Genotype_SE = as.numeric(STD_Genotype_SE),
         across(where(is.numeric), ~ round(., 3))) %>%
  filter(op=="~") %>%
  mutate(fdr_pval = p.adjust(p_value, method = "fdr")) %>%
  select(lhs,model,STD_All,fdr_pval) %>%
  `colnames<-`(c("outcome","model","STD_All","fdr_pval")) %>%
  mutate(model = "Differentiation")

#total - bifactor
estsT_bi_all <- tot_bi_all %>% 
  mutate(Unstand_SE = as.numeric(Unstand_SE),
         STD_Genotype_SE = as.numeric(STD_Genotype_SE),
         across(where(is.numeric), ~ round(., 3))) %>%
  filter(op=="~") %>%
  mutate(fdr_pval = p.adjust(p_value, method = "fdr")) %>%
  select(lhs,model,STD_All,fdr_pval) %>%
  `colnames<-`(c("outcome","model","STD_All","fdr_pval"))  %>%
  mutate(model = "Total problems")

# merge diff and tot
ests_bi_both <- estsD_bi_all %>%
  full_join(estsT_bi_all) %>%
  mutate(model=as.factor(model))

# create dataframe
df_bi <- data.frame(
  model = rep(c("Total problems", "Differentiation"), each = 1),
  outcome = rep(c("p"), times = 2))

# create table hierarchical factor model
tbl_bi_both <- ests_bi_both %>%
  left_join(df_bi, by = c("model", "outcome")) %>%
  mutate(across(where(is.numeric), round, digits=2)) %>%
  select(model,outcome,STD_All,fdr_pval) %>%
  mutate(model = if_else(model == "Differentiation", "d", "t")) %>%
  pivot_wider(names_from = model, 
              values_from = c(STD_All,fdr_pval),
              names_sep = ".") %>%
  select(outcome, ends_with(".d"), ends_with(".t")) %>%
  mutate(outcome=as.factor(outcome))

levels(tbl_bi_both$outcome)[levels(tbl_bi_both$outcome)=="p"] <- "P-factor"
levels(tbl_bi_both$outcome)[levels(tbl_bi_both$outcome)=="com"] <- "Compulsive"
levels(tbl_bi_both$outcome)[levels(tbl_bi_both$outcome)=="psy"] <- "Psychotic"
levels(tbl_bi_both$outcome)[levels(tbl_bi_both$outcome)=="neu"] <- "Neurodevelopmental"
levels(tbl_bi_both$outcome)[levels(tbl_bi_both$outcome)=="int"] <- "Internalising"

(ft_tbl_bi_both <- tbl_bi_both %>% 
    flextable() %>% 
    set_header_labels(values=list(rhs="Latent factor",
                                  STD_All.d="STD_EST",FDR_pval.d="",
                                  STD_All.t="STD_EST",FDR_pval.t="")) %>%
    add_header_row(values=list("","Differentiation","Total problems"), colwidths = c(1,2,2)) %>%
    compose(i = c(2,2), j = c(3,5), part = "header", value = as_paragraph("P",as_sub("FDR"))) %>%
    flextable::font(fontname = "Arial", part= "all") %>% 
    theme_box() %>%
    border_outer(border=fp_border(color="gray80",width=2),part="all") %>%
    border_inner(border=fp_border(color='gray80',width=1),part="all") %>%
    hline(i=2,border=fp_border(color="gray50",width=3),part="header") %>%
    bg(bg="#F1F1F1",part="header") %>% 
    align(align = "center", part="all") %>% 
    set_table_properties(layout = "autofit"))

save_as_docx(ft_tbl_bi_both, path = "./tables/Table_S9_bifactor_model_output.docx")

#plots are generated in the script '04.2_plot_structural_models.R'

#Note abt standardised estimates: STD_All is fully standardised, STD_Genotype_SE 
#is only standardised with respect to the genetic variance in the phenotypes
#This is because, at present, the necessary components to obtain sandwich 
#corrected standard errors are not available for endogenous factors. Use STD_All
